--- bottombar.ios.d.ts
@@ -1,3 +1,0 @@
-export declare class BottomBar {
-    return: any;
-}
