--- bottombar.ios.ts
@@ -1,3 +1,0 @@
-export class BottomBar {
-    return;
-}
