+++ demo/app/app.component.html
@@ -1,12 +1,13 @@
+<ActionBar title="nativescript-bottombar" class="action-bar">
+</ActionBar>
 <GridLayout rows="*, auto">
-    <Label row="0" text="row1"></Label>
+    <StackLayout orientation="vertical">
+        <Button class="btn btn-primary" text="Hide BottomBar" (tap)="hideBottombar()"></Button>
+        <Button class="btn btn-primary" text="Show BottomBar" (tap)="showBottombar()"></Button>
+        <Button class="btn btn-primary" text="Change icon of current tab" (tap)="changeIcon()"></Button>
+        <Button class="btn btn-primary" text="Change color of current tab" (tap)="changeColor()"></Button>
+        <Button class="btn btn-primary" text="Change title of current tab" (tap)="changeTitle()"></Button>
+    </StackLayout>
+    
     <BottomBar #bar row="1" [items]="items" (tabSelected)="tabSelected($event)" [hide]="hidden"></BottomBar>
-    <!-- <BottomBar row="1" (tabSelected)="tabSelected($event)" [hide]="hidden">
-        
-        <BottomBarItem color="#1083BF" title="Home" icon="ic_home_black_24dp" [notification]="notification"></BottomBarItem>
-        <BottomBarItem color="#1083BF" title="Calendar" icon="ic_calendar"></BottomBarItem>
-        <BottomBarItem color="#1083BF" title="Profile" icon="ic_collaborator"></BottomBarItem>
-        <BottomBarItem color="#1083BF" title="Message" icon="ic_paperplane"></BottomBarItem>
-    </BottomBar>-->
-
 </GridLayout>
